# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Publication API",
        "description": "API pour gérer les publications (texte, images et vidéos) les commentaires et les likes",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Server local"
        },
        {
            "url": "/"
        }
    ],
    "paths": {
        "/comments": {
            "post": {
                "tags": [
                    "Comments"
                ],
                "summary": "Créer un nouveau commentaire",
                "operationId": "a0c0613c248bc727d51299e2eb03d2e3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "publication_id",
                                    "content"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "publication_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "Ceci est un commentaire."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Commentaire créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Comment"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erreur de validation"
                    }
                }
            }
        },
        "/comments/{commentId}": {
            "get": {
                "tags": [
                    "Comments"
                ],
                "summary": "Récupérer un commentaire par son ID",
                "operationId": "8398c5ec7573cb377989dc5d4675857e",
                "parameters": [
                    {
                        "name": "commentId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails du commentaire",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Comment"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Commentaire non trouvé"
                    }
                }
            },
            "put": {
                "tags": [
                    "Comments"
                ],
                "summary": "Mettre à jour un commentaire",
                "operationId": "ce01b2e370b2721178467c8af5766380",
                "parameters": [
                    {
                        "name": "commentId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "content"
                                ],
                                "properties": {
                                    "content": {
                                        "type": "string",
                                        "example": "Ceci est un commentaire mis à jour."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Commentaire mis à jour avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Comment"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erreur de validation"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Comments"
                ],
                "summary": "Supprimer un commentaire",
                "operationId": "710124e7bdbfb02f397c1939fb83c90e",
                "parameters": [
                    {
                        "name": "commentId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Commentaire supprimé avec succès"
                    },
                    "400": {
                        "description": "Erreur lors de la suppression"
                    }
                }
            }
        },
        "/publications/{publicationId}/comments": {
            "get": {
                "tags": [
                    "Comments"
                ],
                "summary": "Récupérer tous les commentaires d'une publication",
                "operationId": "446f83b30cdb98e49a99fca6d6d91fc4",
                "parameters": [
                    {
                        "name": "publicationId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des commentaires",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Comment"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Publication non trouvée"
                    }
                }
            }
        },
        "/users/{userId}/comments": {
            "get": {
                "tags": [
                    "Comments"
                ],
                "summary": "Récupérer tous les commentaires d'un utilisateur",
                "operationId": "cd49600ba72a145c485073c1490548e7",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des commentaires",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Comment"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Utilisateur non trouvé"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Comment": {
                "description": "Représente un commentaire",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "content": {
                        "type": "string",
                        "example": "Ceci est un commentaire"
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "publication_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-10-01T12:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-10-01T12:00:00Z"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}